ath79: initial support for RouterStation Pro
This adds the .dts and modifications to base-files / generic-ubnt.mk
necessary to support the Ubiquiti RouterStation Pro on the ath79 target.

All hardware is functional including the AR8216 switch, wireless via
ath9k using the miniPCI slots, flash, USB, button, and LED.

Switch and LAN/WAN configuration is the same as it is with the
equivalent ar71xx target.

Flashing via both sysupgrade and TFTP has been confirmed to work.

Signed-off-by: Matt Merhar <mattmerhar@protonmail.com>