ath79: minor cosmetic fixes to Ubiquiti RouterStation targets
generic-ubnt.mk redundantly specified SUPPORTED_DEVICES for the both the
RouterStation and RouterStation Pro targets. This commit removes that,
so it will instead be auto-calculated by the ath79 Makefile based on the
.dts filenames.

The shared .dtsi only specified the flash model in the corresponding
compatible property, but did not contain the full manufacturer string.
It was also missing the secondary generic "jedec,spi-nor" compatible
property.

Signed-off-by: Matt Merhar <mattmerhar@protonmail.com>